refactor!: Renamed and moved some files
- Capitalized filenames of the subpanels and changed the names of the functions they export.
- Moved pomodoro.tsx into the /subpanel/ folder.
- Renamed todo.tsx to Tasks.tsx.
